Moved Converter tests to Test namespace.
